Reorganize client file structure
