+++ client/src/App.jsx
@@ -1,13 +1,13 @@
 import React, { Component } from 'react';
 import { BrowserRouter as Router, Route} from 'react-router';
 
-import { client } from './redux/reducers';
-import store from './redux/store.js';
+import { client } from './reducers';
+import store from './store.js';
 
-import Nav from './components/Nav.jsx';
+import Nav from './containers/Nav';
 import Search from './layouts/Search.jsx';
 import Project from './layouts/Project.jsx';
-import UserProfile from './layouts/UserProfile.jsx';
+import User from './layouts/User.jsx';
 import Settings from './layouts/Settings.jsx';
 import Login from './layouts/Login.jsx';
 import Signup from './layouts/Signup.jsx';
@@ -36,7 +36,7 @@ const App = () => (
           path='/project/:username/:projectName'
           component={Project}
         />
-        <Route exact path='/user/:username' component={UserProfile}/>
+        <Route exact path='/user/:username' component={User}/>
         <Route exact path='/settings' component={Settings}/>
         <Route exact path='/login' component={Login}/>
         <Route exact path='/signup' component={Signup}/>
