+++ client/src/actions/sessionActions.js
@@ -1,0 +1,6 @@
+const setCurrentUser = (user) => ({
+  type: 'SET_CURRENT_USER',
+  user
+});
+
+export { setCurrentUser };
