+++ client/src/components/MediaComponent/PhotoComponent.jsx
