+++ client/src/components/MediaComponent/TextComponent.jsx
