+++ client/src/index.jsx
@@ -2,10 +2,10 @@ import React from 'react';
 import ReactDOM from 'react-dom';
 import { BrowserRouter } from 'react-router-dom';
 
-import store from './redux/store.js';
+import store from './store.js';
 import './index.css';
 import App from './App.jsx';
-import registerServiceWorker from './registerServiceWorker';
+// import registerServiceWorker from './registerServiceWorker';
 
 ReactDOM.render((
   <BrowserRouter basename="/">
@@ -13,4 +13,4 @@ ReactDOM.render((
   </BrowserRouter>
 ), document.getElementById('root'));
 
-registerServiceWorker.register();
+// registerServiceWorker();
