--- client/src/layouts/Profile.jsx
@@ -1,13 +1,0 @@
-import React, { Component } from 'react';
-
-class Profile extends Component {
-  render() {
-    return (
-      <div>
-        <h2>Profile layout page</h2>
-      </div>
-    );
-  }
-}
-
-export default Profile;
