+++ client/src/layouts/Search.jsx
@@ -1,14 +1,10 @@
 import React, { Component } from 'react';
 import SearchResults from '../containers/SearchResults.jsx';
 
-class Search extends Component {
-  render() {
-    return (
-      <div style={{padding: '2%'}}>
-        <SearchResults />   
-      </div>
-    );
-  }
-}
+const Search = () => (
+  <div style={{padding: '2%'}}>
+    <SearchResults />   
+  </div>
+);
 
 export default Search;
